fix mock update trigger; set status tooltip on whole button
